chore: some more changes
